Toeplitz NUFFT option added to NUFFT linop
- Added normal operators tests
- Relaxed previous tests

+++ tests/test_wavelet.py
@@ -11,14 +11,14 @@ class TestWavelet(unittest.TestCase):
 
     def test_fwt(self):
         n = 8
-        input = np.zeros(n, dtype=np.float)
+        input = np.zeros(n, dtype=np.float32)
         input[0] = 1
         npt.assert_allclose(wavelet.fwt(input, level=1, wave_name='haar'),
                             [1 / 2**0.5, 0, 0, 0, 1 / 2**0.5, 0, 0, 0])
 
     def test_fwt_iwt(self):
         for n in range(5, 11):
-            input = np.zeros(n, dtype=np.float)
+            input = np.zeros(n, dtype=np.float32)
             input[0] = 1
             _, coeff_slices = wavelet.get_wavelet_shape([n])
             npt.assert_allclose(
